+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - Fixed screen reader accessibility issue with the pagination `label` in the `Pagination` component.
 - Fixed focus handling on the expand/collapse icon button in the `ProgressBar` component to address accessibility issues
 - Fixed screen reader accessibility issue with the `label` in `Select` component
+- Added `component` props in `Typography`. 
 - Fixed focus issue for different status icon buttons in the `ProgressBar` component
 
 ### Changed